Texture Generation
Trying to get texture gen working.

Changes to: TileGen: addition of texturearray class, that creates texture array and is passed pre existing colour map, created in buildbiomeTexture.

Changes have also been made to buildbiometexture method: method has been an array, as to return the tiletexture array. This means that the game is not generating for biome, to make it work, make the buildbiome method not an array, and un-comment the existing texture application at the bottom of buiodbiome method, and return the single texture, not the array that is currently being returned.